% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_graph.R
\name{crop_graph}
\alias{crop_graph}
\title{Crop a Spatial Graph to (extent of) Spatial* Object}
\usage{
crop_graph(g, x, delete.islands = F)
}
\arguments{
\item{g}{igraph object with vertex attributes x and y, encoding coordinates.}

\item{x}{Spatial* Object of the same projection as the graph.}

\item{delete.islands}{Delete all disconnected components except the largest one.}
}
\value{
A reduced igraph object.
}
\description{
Function deletes all vertices that are not located within 
x -- if it is a polygon, its area, otherwise its spatial extent.
}
