% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_grid.R
\name{delete_outlier_edges}
\alias{delete_outlier_edges}
\title{Delete edges that are too long on regular grid}
\usage{
delete_outlier_edges(graph, prec = 0.01, max.degree = NULL)
}
\arguments{
\item{graph}{graph}

\item{prec}{precision -- outliers that are longer and still allowed.}

\item{max.degree}{Set maximum for degree. Not used currently.}
}
\value{
A reduced graph
}
\description{
Delete edges that are too long on regular grid
}
