% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatial_graph.R
\name{edges2lines}
\alias{edges2lines}
\title{Transform edges to lines}
\usage{
edges2lines(g)
}
\arguments{
\item{g}{igraph object with vertex attributes x and y, encoding coordinates.}
}
\value{
A four-column matrix of point pairs in the same order as g's edges.
}
\description{
Transforms the graph's edges to plottable lines
}
