% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_intersects.R
\name{edges2sl}
\alias{edges2sl}
\title{Edges to SpatialLines}
\usage{
edges2sl(g)
}
\arguments{
\item{g}{An igraph object with an "x" and "y" coordinate column.}
}
\value{
A SpatialLines object of same length and order as E(g)
}
\description{
Edges to SpatialLines
}
