% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_from.R
\name{graph_from_poly}
\alias{graph_from_poly}
\title{Polygons to spatial graph}
\usage{
graph_from_poly(poly, pts = NULL, plot.result = F)
}
\arguments{
\item{pts}{Optional SpatialPoints* object to set different x and y coordinates for graph vertices.}
}
\value{
An igraph object with an "x" and "y" coordinate attribute and all 
polygon-level data transferred to vertex attributes.
}
\description{
Transforms a set of polygons to a spatial graph. 
Vertices are encodes as polygons centroids (if pts are not supplied) and 
edges are drawn on the basis of the adjacency matrix of polygons (based on \code{sp::gTouches})
}
