% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_from.R
\name{graph_from_pts}
\alias{graph_from_pts}
\title{Points to spatial graph}
\usage{
graph_from_pts(points)
}
\arguments{
\item{points}{SpatialPoints* object}
}
\value{
An igraph object with an "x" and "y" coordinate attribute and all 
point-level data transferred to vertex attributes.
}
\description{
Transforms a set of points to a spatial graph via a Delaunay triangulation.
}
