% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_intersects.R
\name{intersect_lines}
\alias{intersect_lines}
\title{Intersect Lines with Lines}
\usage{
intersect_lines(
  sl,
  sl2,
  aggvar,
  aggfun = function(x) {
     max(x)
 },
  na.val = 0,
  only.uneven = T
)
}
\arguments{
\item{sl}{Line to aggregate}

\item{sl2}{Line that queries}

\item{aggvar}{Variable to aggrgate}

\item{aggfun}{Function for aggregation}

\item{na.val}{NA value}

\item{only.uneven}{only count uneven intersections}
}
\description{
Intersect Lines with Lines
}
