% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_intersects.R
\name{lines2graph}
\alias{lines2graph}
\title{Aggregate SpatialLines* Data to Spatial Graph}
\usage{
lines2graph(sl, graph, aggvar, aggfun, na.val = 0, only.uneven = T)
}
\arguments{
\item{sl}{SpatialLinesDataFrame to aggregate onto spatial graph}

\item{graph}{An igraph object with an "x" and "y" coordinate column.}

\item{aggvar}{Variable in sl to aggregate}

\item{aggfun}{Function to aggregate values if an edge crosses multiple lines.}

\item{na.val}{Value to set if an edge crosses no line}

\item{only.uneven}{Count only intersections where an edge crosses a 
line an uneven number of times. 
This avoids ``fake intersections, where the starts and ends of the
edge are on the same side of the line.}
}
\value{
Vector of length E(g)
}
\description{
Aggregate SpatialLines* Data to Spatial Graph
}
