% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatial_graph.R
\name{plot_spatial_graph}
\alias{plot_spatial_graph}
\title{Plot a spatial graph}
\usage{
plot_spatial_graph(
  g,
  vertex.size = 0.25,
  vertex.color = "black",
  vertex.pch = 19,
  edge.color = "grey",
  edge.width = 0.25,
  axes = F,
  bty = "n",
  xlab = "",
  ylab = "",
  add = F,
  edge.lty = 1,
  ...
)
}
\arguments{
\item{g}{igraph object with vertex attributes x and y, encoding coordinates.}

\item{vertex.size}{Size of vertices}

\item{vertex.color}{Color of vertices}

\item{vertex.pch}{Pch of vertices}

\item{edge.color}{Color of edges}

\item{edge.width}{Width of edges}

\item{axes}{Show axes?}

\item{bty}{Show box?}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{add}{Add to previous plot?}

\item{edge.lty}{Lty of edges}

\item{...}{Additional parameters passed to \code{plot(NULL)}}
}
\value{
Nothing.
}
\description{
Spatial plotting of a graph using R base plot. 
Plots vertices as SpatialPoints and edges as SpatialLines.
}
