% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_intersects.R
\name{poly2graph}
\alias{poly2graph}
\title{Encode Polygon ID on graph}
\usage{
poly2graph(graph, poly, multi = "NA")
}
\arguments{
\item{graph}{An igraph object with an "x" and "y" coordinate column.}

\item{poly}{A SpatialPolygons* object. Function assumes non-overlapping polygons.
Vertices which intersect with two or more polygons, are assigned an NA value.}

\item{multi}{Value to set for intersections with multiple polygons. "NA" = NA, "first" = first, "last" = last}
}
\value{
Vector of length V(graph) which encodes the numberic position of the polygon in poly vertices intersect with.
}
\description{
Encode Polygon ID on graph
}
