% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_intersects.R
\name{raster2graph}
\alias{raster2graph}
\title{Extract raster values to spatial graph}
\usage{
raster2graph(raster, graph, edges = TRUE, aggfun = max, ...)
}
\arguments{
\item{raster}{A raster* object}

\item{graph}{An igraph object with an "x" and "y" coordinate column.}

\item{edges}{Flag on whether to aggregate to edges or vertices.}

\item{aggfun}{Function to aggregate values, when edges = TRUE. Defaults to max.}

\item{...}{Additional parameters passed to \code{raster::extract()}}
}
\value{
The results of raster::extract().
}
\description{
Extract raster values to spatial graph
}
