% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_grid.R
\name{reproject_shape}
\alias{reproject_shape}
\title{Reproject shape for sampling}
\usage{
reproject_shape(shape, proj4string)
}
\arguments{
\item{shape}{Shape}

\item{proj4string}{arget projection}
}
\value{
Reprojected shape
}
\description{
Reproject shape for sampling
}
