% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_grid.R
\name{spatial_grid}
\alias{spatial_grid}
\title{Make a spatial grid}
\usage{
spatial_grid(shape, structure, N_pts, intermediate_proj = NULL)
}
\arguments{
\item{shape}{A shape to cover. SpatialPolygons* object.}

\item{structure}{Sampling and connectivity structure. 
One of: quadratic, hexagonal, triangular, and random. 
Quadratic puts vertices on a regular grid and connects each vertex its direct 4 neighbours. 
Hexagonal samples vertices as the centroids of hexagons and connects each vertex with its direct 6 neighbours. 
Triangular samples vertices as the corners of hexagons and connects each vertex with its direct 3 neighbours. 
Random samples vertices randomly within the shape, connecting each with its neighbours through a Delaunay triangulation.}

\item{N_pts}{Number of points (approximate in structure not random)}

\item{intermediate_proj}{Intermediate projection to use for sampling. 
Can be used to sample from, e.g., an equal area Albers projection that creates spatially more regular grids.}
}
\value{
An igraph object with x and y coordinates encoded as vertex attributes.
}
\description{
Make a spatial grid
}
