% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_grid.R
\name{spatial_hexgrid}
\alias{spatial_hexgrid}
\title{Make grid with hexagonal sampling and connectivity structure}
\usage{
spatial_hexgrid(shape, N_pts, intermediate_proj = NULL)
}
\arguments{
\item{shape}{A shape to cover. SpatialPolygons* object.}

\item{N_pts}{Number of points}

\item{intermediate_proj}{Intermediate projection to use for sampling.}
}
\value{
An igraph object with x and y coordinates encoded as vertex attributes.
}
\description{
Make grid with hexagonal sampling and connectivity structure
}
