% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_graph.R
\name{split_graph}
\alias{split_graph}
\title{Split graph into smaller pieces}
\usage{
split_graph(
  g,
  type = c("sprandom", "spregular", "sphexagonal", "connected_component", "kmeans",
    "byvar"),
  size = NULL,
  drop.smaller.than = 1,
  variable = NULL
)
}
\arguments{
\item{g}{igraph object with vertex attributes x and y, encoding coordinates.}

\item{type}{Type of split applied. 
sprandom, spregular, sphexagonal applies the respective spsample technique to vertices. 
 connected_component splits into connected components. 
 kmeans applies \code{stats::kmeans} clustering to vertices spatial coordinates.
 byvar splits by grouping variable.}

\item{size}{Average size of clusteres required.}

\item{drop.smaller.than}{Drop clusters of size x.}

\item{variable}{If \code{type = byvar}, this is the name of the grouping vertex attribute to split by.}
}
\value{
List of graphs
}
\description{
Split graph into smaller pieces
}
