% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_graph.R
\name{vertex_within}
\alias{vertex_within}
\title{Checks whether vertices are located in (exntent of) Spatial* object}
\usage{
vertex_within(g, x)
}
\arguments{
\item{g}{igraph object with vertex attributes x and y, encoding coordinates.}

\item{x}{Spatial* Object of the same projection as the graph.}
}
\value{
A reduced igraph object.
}
\description{
Function checks for each whether it is located within 
x -- if it is a polygon, its area, otherwise its spatial extent.
}
