% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_intersects.R
\name{vertices2sp}
\alias{vertices2sp}
\title{Vertices to SpatialPoints}
\usage{
vertices2sp(g)
}
\arguments{
\item{g}{An igraph object with an "x" and "y" coordinate column.}
}
\value{
A SpatialPoints object of same length and order as V(g)
}
\description{
Vertices to SpatialPoints
}
