% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_intersects.R
\name{vid_2_edge_diff}
\alias{vid_2_edge_diff}
\title{Binary edge encoding of vertex differences}
\usage{
vid_2_edge_diff(g, attr_name)
}
\arguments{
\item{g}{Graph}

\item{attr_name}{vertex attribute to compute edge-level differences on}
}
\value{
Binary vector of length E(g)
}
\description{
Binary edge encoding of vertex differences
}
